requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ─── Async Queue ──────────────────────────────────────────────────────────────
celery==5.3.6
kombu==5.3.4

# ─── Environment & Config ─────────────────────────────────────────────────────
django-environ==0.11.2

# ─── CORS ─────────────────────────────────────────────────────────────────────
django-cors-headers==4.3.1

# ─── XML Generation (PEPPOL UBL) ──────────────────────────────────────────────
lxml==5.2.1

# ─── PDF Generation (Visual Invoice) ─────────────────────────────────────────
xhtml2pdf==0.2.16

# ─── Utilities ────────────────────────────────────────────────────────────────
Pillow==10.3.0
python-dateutil==2.9.0
uuid==1.30

# ─── Development ──────────────────────────────────────────────────────────────
django-debug-toolbar==4.3.0
ipython==8.24.0
factory-boy==3.3.0
pytest-django==4.8.0
pytest==8.2.0
black==24.4.2
flake8==7.0.0

# ─── Production ───────────────────────────────────────────────────────────────
gunicorn==22.0.0
boto3==1.34.84
django-storages==1.14.3
sentry-sdk==1.45.0
